Added a file for interviewer efficiency
The last two columns tell us about the performance of a candidate after the result of the interview. There are categories for the performance which can be seen as a failure of the interviewer or the success of an interviewer. If a recruited candidate performs below expectations, the interviewer is gives "f". If he performs above expectations, as per expectations or outstanding, the interviewer is given a "p". For candidates who weren't recruited, if they join an inferior competition at a higher level, the interviewer is marked an "f" because there is something that the other company saw which he failed to see. If the candidate joins the inferior competition at a lower level, the interviewer is marked "p". 
If the candidate is promoted in the current organisation, the interviewer is marked "f". If not promoted then a "p" mark is assigned.
Finally, if the rejected candidate joins a superior competition at same level or at a higher level, the interviewer is marked "f".